+++ app/routes/api_config.py
@@ -9,4 +9,3 @@ api_router = APIRouter()
 
 
 api_router.include_router(usuario_routes.router, prefix='/usuarios', tags=['usuarios'])
-api_router.include_router(usuario_routes.router, prefix='/usuarios-google', tags=['usuarios-google'])
